Merge non-VRX addon packages into single `virelex` package
This commit is broadly a marging of the directory hierarchies of the
following single-node and single-interface packages into the new
`virelex` package:

Node packages:
- `action_planner`
- `collision_detector`
- `lidar_processor`
- `state_processor`

Interface packages:
- `cluster_msg`
- `robot_info_msg`
- `state_msg`

A valid argument could be made for splitting the `virelex` package in
two:  `virelex` and `virelex_msgs` (the former having a dependency on
the latter).  However, the message format definitions used here are
designed to be all-but-exclusively used within the scope of the
`virelex` nodes (and their interaction with VRX/Gazebo), so hosting
everything under a single package is reasonable.

The repo at this commit is confirmed to build, but integrated runtime
tests have not been performed (and are likely to fail, given the
additional refinements needed to the `FetchContent` handling of the
`libtorch` dependency from `action_planner`).