Prevent segfault on `action_planner_node` due world name mismatch
Due to the way the gz->ros message bridge is configured for the
simulated sensor suite implemented in VRX, the functionality of
simulated sensor messages being captured is dependent upon the `name`
attribute of the `world` sdf element precisely matching the sdf
filename it is contained within (sans extension).  This commit updates
the generated sdf world file (with buoys added) such that the name
attribute matches the new filename.