Implement alternative experiment runner for arbitrary element spawn
In addition to the title-line implementation of
`virelex/runner_aux.py`, this commit more broadly replaces the
custom collision detection system implemented by
`collision_detector_node.cpp` with the `contact` sensor plugin native
to the Gazebo simulation.
This change is necessary for arbitrary environmental obstacles spawned
in the environment to be considered when evaluating collisions, as the
existing collision detector only recognizes collisions between objects
it is explicitly informed of.

Further, the `unified.launch.xml` file is modified to support obstacle
parameters in list format, which may be used to spawn arbitrary sets
of obstacle objects at runtime, via the `obstacles` launch argument.