Lower `ruff` config python target version to 3.11
This is a workaround for Ubuntu 24.04's `python3-mypy` package;
despite shipping Python 3.12, the `mypy` version is `1.9`, which
doesn't support typechecking for the latest syntax constructs of 3.12.

Lowering the target version to 3.11 prevents `ruff` from telling us to
use the new `type` keyword instead of `TypeAlias`, as well as the new
generic class syntax.

We also modify any uses of these constructs to use the Python
3.11-compatible syntaxes (in `sdflib`'s `__init__.py`).